* Align container ends on 4 grain. 	While this is not very friendly to 64bit architectures, it is 	the only way to allow writing containers portably. It is 	better to always align to an arbitrary value than to align to 	a platform-dependent one.

+++ uexception.h
@@ -138,7 +138,7 @@ public:
     virtual void	write (ostream& os) const;
     virtual size_t	stream_size (void) const;
 protected:
-    int			m_Errno;		///< Error code returned by the failed operation.
+    intptr_t		m_Errno;		///< Error code returned by the failed operation.
     const char*		m_Operation;		///< Name of the failed operation.
 };
